abstracted ensemble experiment structure importer from jingles experiment; minor cleanup
